Related work items: #94307, #95819, #103879, #106071, #106073, #106082, #106100, #106106, #106252, #106259, #106399, #106442, #106524

+++ Logitude/Logitude.BL/CommonDataModel/Tools/EntityService/DocumentsFilingService.cs
@@ -836,9 +836,11 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
             {
                 AddToTasksQueue(theEntityPm, isNewEntity, loggedUserId);
             }
-           
+
             entityRepository.Update(Poco);
             entityRepository.SubmitChanges();
+
+            UpdateCustomsDocumentMetaDataValuesCollection();
 
             AddImporterQueue(theEntityPm, tenantPM, HavingDREL);
             new ShipmentOrderDocumentsQueueService().Build(theEntityPm);
@@ -1038,6 +1040,9 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
             entityRepository.SubmitChanges();
             AddImporterQueue(theEntityPm, tenantPM, HavingDREL);
             new ShipmentOrderDocumentsQueueService().Build(theEntityPm);
+
+            UpdateCustomsDocumentMetaDataValuesCollection();
+
 
             if (entityPM.IsUpdateSharedDocument)
             {
@@ -1780,11 +1785,9 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
 
         private void UpdateDocumentsFilingMetaDataValuesCollection()
         {
-            var isExport = false;
             if (documentsFilingMetaDataValueChangeSet != null)
             {
-                if (this.entityPM.ExternalEntityName == "EFIFILEM" || this.entityPM.ExternalEntityName == "MFIFILEM")
-                    isExport = true;
+             
                 foreach (DocumentsFilingMetaDataValuePM itemPM in documentsFilingMetaDataValueChangeSet)
                 {
                     switch (itemPM.ChangeSetOp)
@@ -1792,27 +1795,21 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
                         case ChangeSetOperation.Insert:
                             {
                                 this.CreateDocumentsFilingMetaDataValue(itemPM);
-                                if(isExport)
-                                    this.CreateCustomsDocumentsFilingMetaDataValue(itemPM);
-
+                              
                                 break;
                             }
 
                         case ChangeSetOperation.Update:
                             {
                                 this.UpdateDocumentsFilingMetaDataValue(itemPM);
-                                if (isExport)
-                                    this.UpdateCustomsDocumentsFilingMetaDataValue(itemPM);
-
+                             
                                 break;
                             }
 
                         case ChangeSetOperation.Delete:
                             {
                                 this.DeleteDocumentsFilingMetaDataValue(itemPM);
-                                if (isExport)
-                                    this.DeleteCustomsDocumentsFilingMetaDataValue(itemPM);
-
+                              
                                 break;
                             }
                       
@@ -1825,6 +1822,47 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
             }
 
            
+
+        }
+        private void UpdateCustomsDocumentMetaDataValuesCollection()
+        {
+           
+            if (documentsFilingMetaDataValueChangeSet != null)
+            {
+                if (this.entityPM.ExternalEntityName == "EFIFILEM" || this.entityPM.ExternalEntityName == "MFIFILEM")
+                {
+
+                    foreach (DocumentsFilingMetaDataValuePM itemPM in documentsFilingMetaDataValueChangeSet)
+                    {
+                        switch (itemPM.ChangeSetOp)
+                        {
+                            case ChangeSetOperation.Insert:
+                                {
+                                     this.CreateCustomsDocumentsFilingMetaDataValue(itemPM);
+                                     break;
+                                }
+
+                            case ChangeSetOperation.Update:
+                                {
+                                       this.UpdateCustomsDocumentsFilingMetaDataValue(itemPM);
+                                       break;
+                                }
+
+                            case ChangeSetOperation.Delete:
+                                {
+                                     this.DeleteCustomsDocumentsFilingMetaDataValue(itemPM);
+                                       break;
+                                }
+
+                            default: { break; }
+                        }
+                    }
+                }
+
+                customContext.SaveChanges();
+            }
+
+
 
         }
 
@@ -1856,23 +1894,35 @@ namespace Logitude.BL.CommonDataModel.Tools.EntityService
 
         private void CreateCustomsDocumentsFilingMetaDataValue(DocumentsFilingMetaDataValuePM itemPM)
         {
-            CustomsDocumentMetaDataValue customsDocumentMetaDataValue = new CustomsDocumentMetaDataValue();
-            customsDocumentMetaDataValue.MetaDataTypeCode = itemPM.DocumentsMetaDataTypeCode;
-            customsDocumentMetaDataValue.CustomsDocumentId = entityPM.Id;
-            customsDocumentMetaDataValue.Tenant = tenant; 
-            customsDocumentMetaDataValue.MetaDataValue = itemPM.MetaDataValue;
-            customsDocumentMetaDataValueRepository.Add(customsDocumentMetaDataValue);
+            CustomsDocumentQueryService myCustomsDocumentQueryService = new CustomsDocumentQueryService(tenant);
+            string documentsFilingId = myCustomsDocumentQueryService.GetDocumentInIdByCustomsDocId(itemPM.Id, tenant);
+
+            if (documentsFilingId != null)
+            {
+                CustomsDocumentMetaDataValue customsDocumentMetaDataValue = new CustomsDocumentMetaDataValue();
+                customsDocumentMetaDataValue.MetaDataTypeCode = itemPM.DocumentsMetaDataTypeCode;
+                customsDocumentMetaDataValue.CustomsDocumentId = entityPM.Id;
+                customsDocumentMetaDataValue.Tenant = tenant;
+                customsDocumentMetaDataValue.MetaDataValue = itemPM.MetaDataValue;
+                customsDocumentMetaDataValueRepository.Add(customsDocumentMetaDataValue);
+            }
         }
         private void UpdateCustomsDocumentsFilingMetaDataValue(DocumentsFilingMetaDataValuePM itemPM)
         {
+
             CustomsDocumentMetaDataValue customsDocumentMetaDataValue = customsDocumentMetaDataValueRepository.GetCustomsDocumentMetaDataValuesByCustomDocumentAndMetaDateValue(entityPM.Id, tenant, itemPM.DocumentsMetaDataTypeCode);
-            customsDocumentMetaDataValue.MetaDataValue=itemPM.MetaDataValue;
-            customsDocumentMetaDataValueRepository.Update(customsDocumentMetaDataValue);
+            if(customsDocumentMetaDataValue != null) {
+                customsDocumentMetaDataValue.MetaDataValue = itemPM.MetaDataValue;
+                customsDocumentMetaDataValueRepository.Update(customsDocumentMetaDataValue);
+            }
+           
         }
         private void DeleteCustomsDocumentsFilingMetaDataValue(DocumentsFilingMetaDataValuePM itemPM)
         {
-            CustomsDocumentMetaDataValue customsDocumentMetaDataValue = customsDocumentMetaDataValueRepository.GetCustomsDocumentMetaDataValuesByCustomDocumentAndMetaDateValue(entityPM.Id, tenant, itemPM.DocumentsMetaDataTypeCode);
-            customsDocumentMetaDataValueRepository.Remove(customsDocumentMetaDataValue);
+
+            CustomsDocumentMetaDataValue customsDocumentMetaDataValue = customsDocumentMetaDataValueRepository.GetCustomsDocumentMetaDataValuesByCustomDocumentAndMetaDateValue(entityPM.Id, tenant,itemPM.DocumentsMetaDataTypeCode);
+            if(customsDocumentMetaDataValue != null) 
+               customsDocumentMetaDataValueRepository.Remove(customsDocumentMetaDataValue);
         }
 
         public string RandomString(int length)
